add date logic to pick right Tikkie
This adds some small logic to pick the right Tikkie from
a list of Tikkies, so we can prefill our payment link with
about two weeks of valid Tikkies, and we don't have to
manually update it for every event.

Important to note: it picks either
- the earliest Tikkie if that is for a date in the future
- the latest Tikkie if that is for a date in the past
- the Tikkie from today (if it exists) if it is after 12:00
- the last Tikkie before today if it is before 12:00

You can optionally add the 'valid-until' date so we can show
a warning if the Tikkie is not valid anymore (also good for our-
selves to check if we still have valid Tikkies).

Important: it will always pick _some_ Tikkie as long as there is
one!